fix(transformer): fixed github sync issue (#7)
I copied the updated files to a local folder,, re-synced with main (which was out of date), started a new branch, then copied the files back in, committed them, and am now merging the correctly updated files with main. I will then delete this branch and continue with everything correctly synced.